+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `ui/Transition` prop `onShow` that fires when transitioning into view a component.
+
 ## [1.1.0] - 2017-04-21
 
 ### Changed